A whole bunch of changes
The email address was updated to the president's address.
The initial sample post was deleted and replaced with first.md
Added a contact page containing a Google Map to EB2
Added a link to sponors on About page
Added a link to calendar as part of config.yml for easier revision later, if needed.